Organize integration tests
