+++ pom.xml
@@ -126,6 +126,39 @@
         </extensions>
         <plugins>
             <plugin>
+                <groupId>org.codehaus.mojo</groupId>
+                <artifactId>build-helper-maven-plugin</artifactId>
+                <version>3.0.0</version>
+                <executions>
+                    <execution>
+                        <id>add-integration-test-source</id>
+                        <phase>generate-test-sources</phase>
+                        <goals>
+                            <goal>add-test-source</goal>
+                        </goals>
+                        <configuration>
+                            <sources>
+                                <source>src/integrationtests/java</source>
+                            </sources>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>add-integration-test-resource</id>
+                        <phase>generate-test-resources</phase>
+                        <goals>
+                            <goal>add-test-resource</goal>
+                        </goals>
+                        <configuration>
+                            <resources>
+                                <resource>
+                                    <directory>src/integrationtests/resources</directory>
+                                </resource>
+                            </resources>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>3.0.0-M4</version>
                 <configuration>
@@ -171,7 +204,9 @@
                         </goals>
                         <configuration>
                             <skipTests>false</skipTests>
-                            <excludedGroups>Integration</excludedGroups>
+                            <excludes>
+                                <exclude>**/integrationtests/**</exclude>
+                            </excludes>
                         </configuration>
                     </execution>
 
@@ -183,8 +218,9 @@
                         </goals>
                         <configuration>
                             <skipTests>false</skipTests>
-                            <excludedGroups>*</excludedGroups>
-                            <groups>Integration</groups>
+                            <includes>
+                                <include>**/integrationtests/**</include>
+                            </includes>
                             <reuseForks>false
                             </reuseForks> <!-- This will make sure to spawn a fresh jvm for each test -->
                         </configuration>
