+++ src/integrationtests/java/com/aws/iot/evergreen/integrationtests/ipc/IPCAwareServicesTest.java
@@ -1,17 +1,23 @@
-package com.aws.iot.evergreen.ipc;
+/* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0 */
+
+package com.aws.iot.evergreen.integrationtests.ipc;
 
 import com.aws.iot.evergreen.dependency.State;
 import com.aws.iot.evergreen.kernel.EvergreenService;
 import com.aws.iot.evergreen.kernel.Kernel;
+import com.aws.iot.evergreen.testcommons.extensions.PerformanceReporting;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+@ExtendWith(PerformanceReporting.class)
 public class IPCAwareServicesTest {
 
     public static Kernel kernel;
@@ -21,7 +27,7 @@ public class IPCAwareServicesTest {
         kernel = new Kernel();
         String tdir = System.getProperty("user.dir");
         kernel.parseArgs("-r", tdir, "-log", "stdout", "-i",
-                IPCServicesTest.class.getResource("ipc_aware_main.yaml").toString());
+                IPCAwareServicesTest.class.getResource("ipc_aware_main.yaml").toString());
         kernel.launch();
     }
