+++ src/integrationtests/java/com/aws/iot/evergreen/integrationtests/kernel/KernelShutdownTest.java
@@ -1,10 +1,15 @@
-package com.aws.iot.evergreen.kernel;
+package com.aws.iot.evergreen.integrationtests.kernel;
 
 import com.aws.iot.evergreen.config.Topic;
 import com.aws.iot.evergreen.config.WhatHappened;
 import com.aws.iot.evergreen.dependency.State;
+import com.aws.iot.evergreen.kernel.EvergreenService;
+import com.aws.iot.evergreen.kernel.GenericExternalService;
+import com.aws.iot.evergreen.kernel.Kernel;
+import com.aws.iot.evergreen.testcommons.extensions.PerformanceReporting;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.util.Collections;
 import java.util.List;
@@ -15,6 +20,7 @@ import java.util.stream.Collectors;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+@ExtendWith(PerformanceReporting.class)
 public class KernelShutdownTest {
 
     private static Kernel kernel;
