+++ src/integrationtests/java/com/aws/iot/evergreen/integrationtests/util/ExecTest.java
@@ -1,10 +1,10 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0 */
 
-package com.aws.iot.evergreen.util;
+package com.aws.iot.evergreen.integrationtests.util;
 
-import com.aws.iot.evergreen.extension.PerformanceReporting;
-import org.junit.jupiter.api.Tag;
+import com.aws.iot.evergreen.testcommons.extensions.PerformanceReporting;
+import com.aws.iot.evergreen.util.Exec;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -18,7 +18,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(PerformanceReporting.class)
-@Tag("Integration")
 public class ExecTest {
     @Test
     public void test() {
