+++ src/integrationtests/resources/com/aws/iot/evergreen/integrationtests/ipc/ipc.yaml
