+++ src/integrationtests/resources/com/aws/iot/evergreen/integrationtests/ipc/ipc_aware_main.yaml
