+++ src/integrationtests/resources/com/aws/iot/evergreen/integrationtests/kernel/delta.yaml
