+++ src/integrationtests/resources/com/aws/iot/evergreen/integrationtests/kernel/long_running_services.yaml
