+++ src/integrationtests/resources/com/aws/iot/evergreen/integrationtests/kernel/single_service.yaml
