+++ src/test/java/com/aws/iot/evergreen/config/PlatformResolverTest.java
@@ -1,3 +1,6 @@
+/* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0 */
+
 package com.aws.iot.evergreen.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
