+++ src/test/java/com/aws/iot/evergreen/ipc/AuthHandlerTest.java
@@ -1,3 +1,6 @@
+/* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0 */
+
 package com.aws.iot.evergreen.ipc;
 
 import com.aws.iot.evergreen.config.Configuration;
