+++ src/test/java/com/aws/iot/evergreen/packagemanager/plugins/impl/LocalPackageStoreTests.java
@@ -1,3 +1,6 @@
+/* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0 */
+
 package com.aws.iot.evergreen.packagemanager.plugins.impl;
 
 import com.aws.iot.evergreen.packagemanager.TestHelper;
