+++ src/test/java/com/aws/iot/evergreen/testcommons/extensions/InitTestSuite.java
@@ -1,7 +1,7 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0 */
 
-package com.aws.iot.evergreen.extension;
+package com.aws.iot.evergreen.testcommons.extensions;
 
 import org.junit.jupiter.api.Test;
