+++ src/test/java/com/aws/iot/evergreen/testcommons/extensions/PerformanceReporting.java
@@ -1,7 +1,7 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0 */
 
-package com.aws.iot.evergreen.extension;
+package com.aws.iot.evergreen.testcommons.extensions;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SequenceWriter;
